move remaining gas funcs
